[WIP] Moves towards radicle plugin ui tests
