+++ src/test/kotlin/network/radicle/jetbrains/radiclejetbrainsplugin/pages/ActionMenuFixture.kt
@@ -1,6 +1,6 @@
 // Copyright 2000-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 
-package network.radicle.jetbrains.radiclejetbrainsplugin
+package network.radicle.jetbrains.radiclejetbrainsplugin.pages
 
 import com.intellij.remoterobot.RemoteRobot
 import com.intellij.remoterobot.data.RemoteComponent
